refactor: move table implementation from expression to table file.
