+++ packages/serverpod/test/database/expressions/expression_test.dart
@@ -1,5 +1,6 @@
 import 'package:serverpod/src/database/columns.dart';
 import 'package:serverpod/src/database/expressions.dart';
+import 'package:serverpod/src/database/table.dart';
 import 'package:test/test.dart';
 
 void main() {
